+++ app/src/main/res/layout/fragment_email_auth_screen.xml
@@ -68,6 +68,7 @@
         android:layout_height="wrap_content"
         android:layout_marginTop="20dp"
         android:hint="Confirm Password"
+        app:passwordToggleEnabled="true"
         app:layout_constraintEnd_toEndOf="@+id/nameLayout"
         app:layout_constraintStart_toStartOf="@+id/nameLayout"
         app:layout_constraintTop_toBottomOf="@+id/messageLayout">
@@ -76,7 +77,7 @@
             android:id="@+id/editTextConfirmPassword"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:inputType="text"
+            android:inputType="textPassword"
             android:maxLines="1"
             android:textSize="15sp" />
     </com.google.android.material.textfield.TextInputLayout>
@@ -88,6 +89,7 @@
         android:layout_height="wrap_content"
         android:layout_marginTop="20dp"
         android:hint="Password"
+        app:passwordToggleEnabled="true"
         app:layout_constraintEnd_toEndOf="@+id/nameLayout"
         app:layout_constraintStart_toStartOf="@+id/nameLayout"
         app:layout_constraintTop_toBottomOf="@+id/emailLayout">
@@ -97,7 +99,7 @@
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:inputType="textPassword"
-            android:maxLines="5"
+            android:maxLines="1"
             android:textSize="15sp" />
     </com.google.android.material.textfield.TextInputLayout>
