+++ app/src/main/res/layout/fragment_log_in.xml
@@ -27,6 +27,7 @@
         android:layout_height="wrap_content"
         android:layout_marginTop="20dp"
         android:hint="Password"
+        app:passwordToggleEnabled="true"
         app:layout_constraintEnd_toEndOf="@+id/emailLayout"
         app:layout_constraintStart_toStartOf="@+id/emailLayout"
         app:layout_constraintTop_toBottomOf="@+id/emailLayout">
